api/requirements.txt
alembic==1.13.3
asgiref==3.8.1
blinker==1.8.2
cachetools==5.5.0
certifi==2024.8.30
charset-normalizer==3.4.0
click==8.1.7
cssselect==1.2.0
cssutils==2.11.1
dlib==19.24.6
face-recognition==1.3.0
face_recognition_models==0.3.0
Flask==3.0.3
Flask-Cors==5.0.0
Flask-HTTPAuth==4.8.0
Flask-Migrate==4.0.7
Flask-SQLAlchemy==3.1.1
greenlet==3.1.1
idna==3.10
importlib_metadata==8.5.0
itsdangerous==2.2.0
Jinja2==3.1.4
lxml==5.3.0
Mako==1.3.5
MarkupSafe==2.1.5
more-itertools==10.5.0
mysql-connector-python==9.0.0
numpy==2.0.2
opencv-python==4.10.0.84
pillow==11.0.0
premailer==3.10.0
PyJWT==2.9.0
python-dotenv==1.0.1
requests==2.32.3
scipy==1.13.1
SQLAlchemy==2.0.35
typing_extensions==4.12.2
urllib3==2.2.3
Werkzeug==3.0.4
yagmail==0.15.293
zipp==3.20.2
